+++ .idea/misc.xml
@@ -1,5 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
+  <component name="DesignSurface">
+    <option name="filePathToZoomLevelMap">
+      <map>
+        <entry key="..\:/Users/elahe/AndroidStudioProjects/ColorChecker/app/src/main/res/layout/analysis_fragment.xml" value="0.30978260869565216" />
+        <entry key="..\:/Users/elahe/AndroidStudioProjects/ColorChecker/app/src/main/res/layout/content_main.xml" value="0.30978260869565216" />
+        <entry key="..\:/Users/elahe/AndroidStudioProjects/ColorChecker/app/src/main/res/layout/fragment_first.xml" value="0.31354166666666666" />
+        <entry key="..\:/Users/elahe/AndroidStudioProjects/ColorChecker/app/src/main/res/menu/menu_main.xml" value="0.31354166666666666" />
+        <entry key="..\:/Users/elahe/AndroidStudioProjects/ColorChecker/app/src/main/res/mipmap-anydpi-v26/ic_launcher_round.xml" value="0.3165" />
+      </map>
+    </option>
+  </component>
   <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
