+++ app/src/main/java/com/example/colorchecker/AnalysisFragment.kt
@@ -1,0 +1,38 @@
+package com.example.colorchecker
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.colorchecker.databinding.AnalysisFragmentBinding
+
+class AnalysisFragment: Fragment() {
+
+    private var _binding: AnalysisFragmentBinding? = null
+
+    private val binding get() = _binding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = AnalysisFragmentBinding.inflate(inflater, container, false)
+        val view = binding?.root
+        return view
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
